y-axis as rotation axis; prototype

+++ May 25/CommonStruct.h
@@ -51,7 +51,7 @@ struct ConstantBuffer_Shadowmap
 struct GeometryPointers
 {
 	SimpleVertex*           pVertexPointer = nullptr;
-	WORD*					pIndexPointer = nullptr;
+	UINT*					pIndexPointer = nullptr;
 };
